@@ -1,92 +1,93 @@
 {
   "name": "ant-design-pro",
-  "version": "1.3.0",
+  "version": "2.0.0",
   "description": "An out-of-box UI solution for enterprise applications",
   "private": true,
   "scripts": {
     "precommit": "npm run lint-staged",
-    "start": "cross-env ESLINT=none roadhog dev",
-    "start:no-proxy": "cross-env NO_PROXY=true ESLINT=none roadhog dev",
-    "build": "cross-env ESLINT=none roadhog build",
-    "site": "roadhog-api-doc static && gh-pages -d dist",
-    "analyze": "cross-env ANALYZE=true roadhog build",
+    "presite": "npm run generate-mock && cd functions && npm install",
+    "start": "cross-env APP_TYPE=site umi dev",
+    "start:no-mock": "cross-env MOCK=none umi dev",
+    "build": "umi build",
+    "site": "npm run presite && cross-env APP_TYPE=site npm run build && firebase deploy",
+    "analyze": "cross-env ANALYZE=1 umi build",
     "lint:style": "stylelint \"src/**/*.less\" --syntax less",
     "lint": "eslint --ext .js src mock tests && npm run lint:style",
     "lint:fix": "eslint --fix --ext .js src mock tests && npm run lint:style",
     "lint-staged": "lint-staged",
     "lint-staged:js": "eslint --ext .js",
-    "test": "roadhog test",
-    "test:component": "roadhog test ./src/components",
+    "test": "umi test",
+    "generate-mock": "node ./scripts/generateMock.js",
+    "test:component": "umi test ./src/components",
     "test:all": "node ./tests/run-tests.js",
     "prettier": "prettier --write ./src/**/**/**/*"
   },
   "dependencies": {
-    "@antv/data-set": "^0.8.0",
-    "@babel/polyfill": "^7.0.0-beta.36",
-    "antd": "3.7.3",
+    "@antv/data-set": "^0.9.5",
+    "@babel/runtime": "^7.0.0",
+    "antd": "^3.9.1",
     "bizcharts": "^3.1.10",
-    "bizcharts-plugin-slider": "^2.0.1",
-    "classnames": "^2.2.5",
-    "dva": "^2.2.3",
-    "dva-loading": "^2.0.3",
+    "bizcharts-plugin-slider": "^2.0.3",
+    "classnames": "^2.2.6",
+    "dva": "^2.4.0",
     "echarts": "^4.1.0",
     "enquire-js": "^0.2.1",
+    "hash.js": "^1.1.5",
     "lodash": "^4.17.10",
     "lodash-decorators": "^6.0.0",
-    "moment": "^2.19.3",
+    "memoize-one": "^4.0.0",
+    "moment": "^2.22.2",
     "numeral": "^2.0.6",
+    "nzh": "^1.0.3",
     "omit.js": "^1.0.0",
-    "path-to-regexp": "^2.1.0",
+    "path-to-regexp": "^2.4.0",
     "prop-types": "^15.5.10",
-    "qs": "^6.5.0",
-    "rc-drawer": "^1.2.0",
+    "qs": "^6.5.2",
+    "rc-animate": "^2.4.4",
     "react": "^16.4.1",
     "react-bmap": "^1.0.80",
     "react-container-query": "^0.11.0",
+    "react-copy-to-clipboard": "^5.0.1",
     "react-document-title": "^2.0.3",
-    "react-dom": "^16.4.1",
+    "react-dom": "^16.4.2",
     "react-fittext": "^1.0.0",
     "react-markdown": "^3.4.1",
     "recharts": "^1.0.0-beta.10",
-    "rollbar": "^2.3.4",
-    "setprototypeof": "^1.1.0",
-    "url-polyfill": "^1.0.10"
+    "react-router-config": "^1.0.0-beta.4",
+    "react-router-dom": "^4.3.1"
   },
   "devDependencies": {
-    "babel-eslint": "^8.1.2",
-    "babel-plugin-dva-hmr": "^0.4.1",
-    "babel-plugin-import": "^1.6.7",
-    "babel-plugin-module-resolver": "^3.1.1",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "@types/react": "^16.4.11",
+    "@types/react-dom": "^16.0.6",
+    "antd-pro-merge-less": "^0.0.9",
+    "antd-pro-theme-webpack-plugin": "^1.1.8",
+    "babel-eslint": "^9.0.0",
     "cross-env": "^5.1.1",
     "cross-port-killer": "^1.0.1",
-    "enzyme": "^3.1.0",
-    "eslint": "^5.0.0",
+    "enzyme": "^3.4.4",
+    "eslint": "^5.4.0",
     "eslint-config-airbnb": "^17.0.0",
-    "eslint-config-prettier": "^2.9.0",
+    "eslint-config-prettier": "^3.0.1",
     "eslint-plugin-babel": "^5.1.0",
-    "eslint-plugin-compat": "^2.1.0",
+    "eslint-plugin-compat": "^2.5.1",
     "eslint-plugin-import": "^2.8.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-markdown": "^1.0.0-beta.6",
-    "eslint-plugin-react": "^7.0.1",
-    "gh-pages": "^1.0.0",
-    "husky": "^0.14.3",
+    "eslint-plugin-react": "^7.11.1",
+    "gh-pages": "^1.2.0",
     "lint-staged": "^7.2.0",
+    "merge-umi-mock-data": "^0.0.3",
     "mockjs": "^1.0.1-beta3",
-    "prettier": "1.14.0",
+    "prettier": "1.14.2",
     "pro-download": "^1.0.1",
-    "react-loadable": "^5.4.0",
-    "redbox-react": "^1.5.0",
-    "regenerator-runtime": "^0.12.0",
-    "roadhog": "^2.4.2",
-    "roadhog-api-doc": "^1.1.2",
-    "stylelint": "^9.2.1",
-    "stylelint-config-prettier": "^3.0.4",
-    "stylelint-config-standard": "^18.0.0"
+    "stylelint": "^9.4.0",
+    "stylelint-config-prettier": "^4.0.0",
+    "stylelint-config-standard": "^18.0.0",
+    "umi": "^2.0.1",
+    "umi-plugin-react": "^1.0.1"
   },
   "optionalDependencies": {
-    "puppeteer": "^1.4.0"
+    "puppeteer": "^1.6.0"
   },
   "lint-staged": {
     "**/*.{js,jsx,less}": [
