@@ -10,7 +10,7 @@ export default class PromiseRender extends React.PureComponent {
     this.setRenderComponent(this.props);
   }
 
-  componentWillReceiveProps(nextProps) {
+  componentDidUpdate(nextProps) {
     // new Props enter
     this.setRenderComponent(nextProps);
   }
@@ -45,8 +45,9 @@ export default class PromiseRender extends React.PureComponent {
 
   render() {
     const { component: Component } = this.state;
+    const { ok, error, promise, ...rest } = this.props;
     return Component ? (
-      <Component {...this.props} />
+      <Component {...rest} />
     ) : (
       <div
         style={{
