@@ -5,20 +5,21 @@ import autoHeight from '../autoHeight';
 import styles from './index.less';
 
 /* eslint react/no-danger:0 */
+export default
 @autoHeight()
-export default class Radar extends Component {
+class Radar extends Component {
   state = {
     legendData: [],
   };
 
   componentDidMount() {
-    this.getLengendData();
+    this.getLegendData();
   }
 
-  componentWillReceiveProps(nextProps) {
+  componentDidUpdate(preProps) {
     const { data } = this.props;
-    if (data !== nextProps.data) {
-      this.getLengendData();
+    if (data !== preProps.data) {
+      this.getLegendData();
     }
   }
 
@@ -27,9 +28,10 @@ export default class Radar extends Component {
   };
 
   // for custom lengend view
-  getLengendData = () => {
+  getLegendData = () => {
     if (!this.chart) return;
     const geom = this.chart.getAllGeoms()[0]; // 获取所有的图形
+    if (!geom) return;
     const items = geom.get('dataArray') || []; // 获取图形对应的
 
     const legendData = items.map(item => {
