@@ -3,6 +3,24 @@ import React, { Component } from 'react';
 function fixedZero(val) {
   return val * 1 < 10 ? `0${val}` : val;
 }
+const initTime = props => {
+  let lastTime = 0;
+  let targetTime = 0;
+  try {
+    if (Object.prototype.toString.call(props.target) === '[object Date]') {
+      targetTime = props.target.getTime();
+    } else {
+      targetTime = new Date(props.target).getTime();
+    }
+  } catch (e) {
+    throw new Error('invalid target prop', e);
+  }
+
+  lastTime = targetTime - new Date().getTime();
+  return {
+    lastTime: lastTime < 0 ? 0 : lastTime,
+  };
+};
 
 class CountDown extends Component {
   timer = 0;
@@ -11,56 +29,37 @@ class CountDown extends Component {
 
   constructor(props) {
     super(props);
-
-    const { lastTime } = this.initTime(props);
-
+    const { lastTime } = initTime(props);
     this.state = {
       lastTime,
     };
+  }
+
+  static getDerivedStateFromProps(nextProps, preState) {
+    const { lastTime } = initTime(nextProps);
+    if (preState.lastTime !== lastTime) {
+      return {
+        lastTime,
+      };
+    }
+    return null;
   }
 
   componentDidMount() {
     this.tick();
   }
 
-  componentWillReceiveProps(nextProps) {
+  componentDidUpdate(prevProps) {
     const { target } = this.props;
-    if (target !== nextProps.target) {
+    if (target !== prevProps.target) {
       clearTimeout(this.timer);
-      const { lastTime } = this.initTime(nextProps);
-      this.setState(
-        {
-          lastTime,
-        },
-        () => {
-          this.tick();
-        }
-      );
+      this.tick();
     }
   }
 
   componentWillUnmount() {
     clearTimeout(this.timer);
   }
-
-  initTime = props => {
-    let lastTime = 0;
-    let targetTime = 0;
-    try {
-      if (Object.prototype.toString.call(props.target) === '[object Date]') {
-        targetTime = props.target.getTime();
-      } else {
-        targetTime = new Date(props.target).getTime();
-      }
-    } catch (e) {
-      throw new Error('invalid target prop', e);
-    }
-
-    lastTime = targetTime - new Date().getTime();
-    return {
-      lastTime: lastTime < 0 ? 0 : lastTime,
-    };
-  };
 
   // defaultFormat = time => (
   //  <span>{moment(time).format('hh:mm:ss')}</span>
@@ -74,11 +73,7 @@ class CountDown extends Component {
     const s = Math.floor((time - h * hours - m * minutes) / 1000);
     return (
       <span>
-        {fixedZero(h)}
-        :
-        {fixedZero(m)}
-        :
-        {fixedZero(s)}
+        {fixedZero(h)}:{fixedZero(m)}:{fixedZero(s)}
       </span>
     );
   };
