@@ -13,16 +13,14 @@ const TooltipOverlayStyle = {
   wordWrap: 'break-word',
 };
 
-export const getStrFullLength = (str = '') => {
-  return str.split('').reduce((pre, cur) => {
+export const getStrFullLength = (str = '') =>
+  str.split('').reduce((pre, cur) => {
     const charCode = cur.charCodeAt(0);
     if (charCode >= 0 && charCode <= 128) {
       return pre + 1;
-    } else {
-      return pre + 2;
     }
+    return pre + 2;
   }, 0);
-};
 
 export const cutStrByFullLength = (str = '', maxLength) => {
   let showLength = 0;
@@ -35,9 +33,8 @@ export const cutStrByFullLength = (str = '', maxLength) => {
     }
     if (showLength <= maxLength) {
       return pre + cur;
-    } else {
-      return pre;
     }
+    return pre;
   }, '');
 };
 
@@ -88,9 +85,9 @@ export default class Ellipsis extends Component {
     }
   }
 
-  componentWillReceiveProps(nextProps) {
+  componentDidUpdate(perProps) {
     const { lines } = this.props;
-    if (lines !== nextProps.lines) {
+    if (lines !== perProps.lines) {
       this.computeLine();
     }
   }
@@ -139,29 +136,26 @@ export default class Ellipsis extends Component {
       sh = shadowNode.offsetHeight;
       if (sh > th || mid === begin) {
         return mid;
-      } else {
-        begin = mid;
-        if (end - begin === 1) {
-          mid = 1 + begin;
-        } else {
-          mid = Math.floor((end - begin) / 2) + begin;
-        }
-        return this.bisection(th, mid, begin, end, text, shadowNode);
       }
-    } else {
-      if (mid - 1 < 0) {
-        return mid;
-      }
-      shadowNode.innerHTML = text.substring(0, mid - 1) + suffix;
-      sh = shadowNode.offsetHeight;
-      if (sh <= th) {
-        return mid - 1;
+      begin = mid;
+      if (end - begin === 1) {
+        mid = 1 + begin;
       } else {
-        end = mid;
         mid = Math.floor((end - begin) / 2) + begin;
-        return this.bisection(th, mid, begin, end, text, shadowNode);
       }
+      return this.bisection(th, mid, begin, end, text, shadowNode);
     }
+    if (mid - 1 < 0) {
+      return mid;
+    }
+    shadowNode.innerHTML = text.substring(0, mid - 1) + suffix;
+    sh = shadowNode.offsetHeight;
+    if (sh <= th) {
+      return mid - 1;
+    }
+    end = mid;
+    mid = Math.floor((end - begin) / 2) + begin;
+    return this.bisection(th, mid, begin, end, text, shadowNode);
   };
 
   handleRoot = n => {
