@@ -15,6 +15,10 @@ const TagSelectOption = ({ children, checked, onChange, value }) => (
 TagSelectOption.isTagSelectOption = true;
 
 class TagSelect extends Component {
+  static defaultProps = {
+    hideCheckAll: false,
+  };
+
   constructor(props) {
     super(props);
     this.state = {
@@ -23,10 +27,11 @@ class TagSelect extends Component {
     };
   }
 
-  componentWillReceiveProps(nextProps) {
+  static getDerivedStateFromProps(nextProps) {
     if ('value' in nextProps && nextProps.value) {
-      this.setState({ value: nextProps.value });
+      return { value: nextProps.value };
     }
+    return null;
   }
 
   onChange = value => {
@@ -57,8 +62,8 @@ class TagSelect extends Component {
   }
 
   handleTagChange = (value, checked) => {
-    const { value: v } = this.state;
-    const checkedTags = [...v];
+    const { value: StateValue } = this.state;
+    const checkedTags = [...StateValue];
 
     const index = checkedTags.indexOf(value);
     if (checked && index === -1) {
@@ -76,17 +81,14 @@ class TagSelect extends Component {
     });
   };
 
-  isTagSelectOption = node => {
-    return (
-      node &&
-      node.type &&
-      (node.type.isTagSelectOption || node.type.displayName === 'TagSelectOption')
-    );
-  };
+  isTagSelectOption = node =>
+    node &&
+    node.type &&
+    (node.type.isTagSelectOption || node.type.displayName === 'TagSelectOption');
 
   render() {
     const { value, expand } = this.state;
-    const { children, className, style, expandable } = this.props;
+    const { children, hideCheckAll, className, style, expandable } = this.props;
 
     const checkedAll = this.getAllTags().length === value.length;
 
@@ -96,9 +98,11 @@ class TagSelect extends Component {
     });
     return (
       <div className={cls} style={style}>
-        <CheckableTag checked={checkedAll} key="tag-select-__all__" onChange={this.onSelectAll}>
-          全部
-        </CheckableTag>
+        {hideCheckAll ? null : (
+          <CheckableTag checked={checkedAll} key="tag-select-__all__" onChange={this.onSelectAll}>
+            全部
+          </CheckableTag>
+        )}
         {value &&
           React.Children.map(children, child => {
             if (this.isTagSelectOption(child)) {
