@@ -98,15 +98,15 @@ const salesTypeDataOffline = [
     y: 99,
   },
   {
-    x: '个护健康',
+    x: '食用酒水',
     y: 188,
   },
   {
-    x: '服饰箱包',
+    x: '个护健康',
     y: 344,
   },
   {
-    x: '母婴产品',
+    x: '服饰箱包',
     y: 255,
   },
   {
@@ -118,7 +118,7 @@ const salesTypeDataOffline = [
 const offlineData = [];
 for (let i = 0; i < 10; i += 1) {
   offlineData.push({
-    name: `门店${i}`,
+    name: `Stores ${i}`,
     cvr: Math.ceil(Math.random() * 9) / 10,
   });
 }
@@ -179,7 +179,7 @@ radarOriginData.forEach(item => {
   });
 });
 
-export const getFakeChartData = {
+const getFakeChartData = {
   visitData,
   visitData2,
   salesData,
@@ -193,5 +193,5 @@ export const getFakeChartData = {
 };
 
 export default {
-  getFakeChartData,
+  'GET /api/fake_chart_data': getFakeChartData,
 };
