@@ -4,23 +4,21 @@ let CURRENT = 'NULL';
  * use  authority or getAuthority
  * @param {string|()=>String} currentAuthority
  */
-const renderAuthorize = Authorized => {
-  return currentAuthority => {
-    if (currentAuthority) {
-      if (currentAuthority.constructor.name === 'Function') {
-        CURRENT = currentAuthority();
-      }
-      if (
-        currentAuthority.constructor.name === 'String' ||
-        currentAuthority.constructor.name === 'Array'
-      ) {
-        CURRENT = currentAuthority;
-      }
-    } else {
-      CURRENT = 'NULL';
+const renderAuthorize = Authorized => currentAuthority => {
+  if (currentAuthority) {
+    if (currentAuthority.constructor.name === 'Function') {
+      CURRENT = currentAuthority();
     }
-    return Authorized;
-  };
+    if (
+      currentAuthority.constructor.name === 'String' ||
+      currentAuthority.constructor.name === 'Array'
+    ) {
+      CURRENT = currentAuthority;
+    }
+  } else {
+    CURRENT = 'NULL';
+  }
+  return Authorized;
 };
 
 export { CURRENT };
