@@ -52,7 +52,7 @@ Shape.registerShape('point', 'pointer', {
 });
 
 @autoHeight()
-export default class Gauge extends React.Component {
+class Gauge extends React.Component {
   render() {
     const {
       title,
@@ -142,15 +142,13 @@ export default class Gauge extends React.Component {
           />
           <Html
             position={['50%', '95%']}
-            html={() => {
-              return `
+            html={() => `
                 <div style="width: 300px;text-align: center;font-size: 12px!important;">
                   <p style="font-size: 14px; color: rgba(0,0,0,0.43);margin: 0;">${title}</p>
                   <p style="font-size: 24px;color: rgba(0,0,0,0.85);margin: 0;">
                     ${data[0].value * 10}%
                   </p>
-                </div>`;
-            }}
+                </div>`}
           />
         </Guide>
         <Geom
@@ -165,3 +163,5 @@ export default class Gauge extends React.Component {
     );
   }
 }
+
+export default Gauge;
