@@ -4,7 +4,7 @@ import autoHeight from '../autoHeight';
 import styles from '../index.less';
 
 @autoHeight()
-export default class MiniArea extends React.Component {
+class MiniArea extends React.PureComponent {
   render() {
     const {
       height,
@@ -104,3 +104,5 @@ export default class MiniArea extends React.Component {
     );
   }
 }
+
+export default MiniArea;
