@@ -1,11 +1,11 @@
 import * as React from 'react';
 export interface ITimelineChartProps {
   data: Array<{
-    x: string;
-    y1: string;
-    y2: string;
+    x: number;
+    y1: number;
+    y2?: number;
   }>;
-  titleMap: { y1: string; y2: string };
+  titleMap: { y1: string; y2?: string };
   padding?: [number, number, number, number];
   height?: number;
   style?: React.CSSProperties;
