@@ -6,7 +6,7 @@ import autoHeight from '../autoHeight';
 import styles from './index.less';
 
 @autoHeight()
-export default class TimelineChart extends React.Component {
+class TimelineChart extends React.Component {
   render() {
     const {
       title,
@@ -120,3 +120,5 @@ export default class TimelineChart extends React.Component {
     );
   }
 }
+
+export default TimelineChart;
