@@ -35,17 +35,25 @@ export default class HeaderSearch extends PureComponent {
     };
   }
 
+  componentWillUnmount() {
+    clearTimeout(this.timeout);
+  }
+
   onKeyDown = e => {
     if (e.key === 'Enter') {
-      this.debouncePressEnter();
+      const { onPressEnter } = this.props;
+      const { value } = this.state;
+      this.timeout = setTimeout(() => {
+        onPressEnter(value); // Fix duplicate onPressEnter
+      }, 0);
     }
   };
 
   onChange = value => {
-    this.setState({ value });
     const { onChange } = this.props;
+    this.setState({ value });
     if (onChange) {
-      onChange();
+      onChange(value);
     }
   };
 
@@ -65,18 +73,18 @@ export default class HeaderSearch extends PureComponent {
     });
   };
 
-  debouncePressEnter() {
-    const { onPressEnter } = this.props;
-    const { value } = this.state;
-    onPressEnter(value);
-  }
-
   // NOTE: 不能小于500，如果长按某键，第一次触发auto repeat的间隔是500ms，小于500会导致触发2次
   @Bind()
   @Debounce(500, {
     leading: true,
     trailing: false,
   })
+  debouncePressEnter() {
+    const { onPressEnter } = this.props;
+    const { value } = this.state;
+    onPressEnter(value);
+  }
+
   render() {
     const { className, placeholder, ...restProps } = this.props;
     const { searchMode, value } = this.state;
@@ -95,10 +103,11 @@ export default class HeaderSearch extends PureComponent {
           onChange={this.onChange}
         >
           <Input
-            placeholder={placeholder}
             ref={node => {
               this.input = node;
             }}
+            aria-label={placeholder}
+            placeholder={placeholder}
             onKeyDown={this.onKeyDown}
             onBlur={this.leaveSearchMode}
           />
