@@ -1,113 +1,145 @@
 import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import { Form, Button, Row, Col } from 'antd';
+import { Form, Input, Button, Row, Col } from 'antd';
 import omit from 'omit.js';
 import styles from './index.less';
-import map from './map';
+import ItemMap from './map';
+import LoginContext from './loginContext';
 
 const FormItem = Form.Item;
 
-function generator({ defaultProps, defaultRules, type }) {
-  return WrappedComponent => {
-    return class BasicComponent extends Component {
-      static contextTypes = {
-        form: PropTypes.object,
-        updateActive: PropTypes.func,
-      };
+class WarpFormItem extends Component {
+  static defaultProps = {
+    buttonText: '获取验证码',
+  };
 
-      constructor(props) {
-        super(props);
-        this.state = {
-          count: 0,
-        };
-      }
+  constructor(props) {
+    super(props);
+    this.state = {
+      count: 0,
+    };
+  }
 
-      componentDidMount() {
-        const { updateActive } = this.context;
-        const { name } = this.props;
-        if (updateActive) {
-          updateActive(name);
-        }
-      }
+  componentDidMount() {
+    const { updateActive, name } = this.props;
+    if (updateActive) {
+      updateActive(name);
+    }
+  }
 
-      componentWillUnmount() {
+  componentWillUnmount() {
+    clearInterval(this.interval);
+  }
+
+  onGetCaptcha = () => {
+    const { onGetCaptcha } = this.props;
+    const result = onGetCaptcha ? onGetCaptcha() : null;
+    if (result === false) {
+      return;
+    }
+    if (result instanceof Promise) {
+      result.then(this.runGetCaptchaCountDown);
+    } else {
+      this.runGetCaptchaCountDown();
+    }
+  };
+
+  getFormItemOptions = ({ onChange, defaultValue, rules }) => {
+    const options = {
+      rules: rules || this.customprops.rules,
+    };
+    if (onChange) {
+      options.onChange = onChange;
+    }
+    if (defaultValue) {
+      options.initialValue = defaultValue;
+    }
+    return options;
+  };
+
+  runGetCaptchaCountDown = () => {
+    const { countDown } = this.props;
+    let count = countDown || 59;
+    this.setState({ count });
+    this.interval = setInterval(() => {
+      count -= 1;
+      this.setState({ count });
+      if (count === 0) {
         clearInterval(this.interval);
       }
-
-      onGetCaptcha = () => {
-        let count = 59;
-        this.setState({ count });
-        const { onGetCaptcha } = this.props;
-        if (onGetCaptcha) {
-          onGetCaptcha();
-        }
-        this.interval = setInterval(() => {
-          count -= 1;
-          this.setState({ count });
-          if (count === 0) {
-            clearInterval(this.interval);
-          }
-        }, 1000);
-      };
-
-      render() {
-        const { form } = this.context;
-        const { getFieldDecorator } = form;
-        const options = {};
-        let otherProps = {};
-        const { onChange, defaultValue, rules, name, ...restProps } = this.props;
-        const { count } = this.state;
-        options.rules = rules || defaultRules;
-        if (onChange) {
-          options.onChange = onChange;
-        }
-        if (defaultValue) {
-          options.initialValue = defaultValue;
-        }
-        otherProps = restProps || otherProps;
-        if (type === 'Captcha') {
-          const inputProps = omit(otherProps, ['onGetCaptcha']);
-          return (
-            <FormItem>
-              <Row gutter={8}>
-                <Col span={16}>
-                  {getFieldDecorator(name, options)(
-                    <WrappedComponent {...defaultProps} {...inputProps} />
-                  )}
-                </Col>
-                <Col span={8}>
-                  <Button
-                    disabled={count}
-                    className={styles.getCaptcha}
-                    size="large"
-                    onClick={this.onGetCaptcha}
-                  >
-                    {count ? `${count} s` : '获取验证码'}
-                  </Button>
-                </Col>
-              </Row>
-            </FormItem>
-          );
-        }
-        return (
-          <FormItem>
-            {getFieldDecorator(name, options)(
-              <WrappedComponent {...defaultProps} {...otherProps} />
-            )}
-          </FormItem>
-        );
-      }
-    };
+    }, 1000);
   };
+
+  render() {
+    const { count } = this.state;
+
+    const {
+      form: { getFieldDecorator },
+    } = this.props;
+
+    // 这么写是为了防止restProps中 带入 onChange, defaultValue, rules props
+    const {
+      onChange,
+      customprops,
+      defaultValue,
+      rules,
+      name,
+      buttonText,
+      updateActive,
+      type,
+      ...restProps
+    } = this.props;
+
+    // get getFieldDecorator props
+    const options = this.getFormItemOptions(this.props);
+
+    const otherProps = restProps || {};
+    if (type === 'Captcha') {
+      const inputProps = omit(otherProps, ['onGetCaptcha', 'countDown']);
+      return (
+        <FormItem>
+          <Row gutter={8}>
+            <Col span={16}>
+              {getFieldDecorator(name, options)(<Input {...customprops} {...inputProps} />)}
+            </Col>
+            <Col span={8}>
+              <Button
+                disabled={count}
+                className={styles.getCaptcha}
+                size="large"
+                onClick={this.onGetCaptcha}
+              >
+                {count ? `${count} s` : buttonText}
+              </Button>
+            </Col>
+          </Row>
+        </FormItem>
+      );
+    }
+    return (
+      <FormItem>
+        {getFieldDecorator(name, options)(<Input {...customprops} {...otherProps} />)}
+      </FormItem>
+    );
+  }
 }
 
 const LoginItem = {};
-Object.keys(map).forEach(item => {
-  LoginItem[item] = generator({
-    defaultProps: map[item].props,
-    defaultRules: map[item].rules,
-    type: item,
-  })(map[item].component);
+Object.keys(ItemMap).forEach(key => {
+  const item = ItemMap[key];
+  LoginItem[key] = props => (
+    <LoginContext.Consumer>
+      {context => (
+        <WarpFormItem
+          customprops={item.props}
+          {...props}
+          rules={item.rules}
+          type={key}
+          updateActive={context.updateActive}
+          form={context.form}
+        />
+      )}
+    </LoginContext.Consumer>
+  );
 });
 
 export default LoginItem;
