@@ -13,8 +13,8 @@ export default function Result({
   ...restProps
 }) {
   const iconMap = {
-    error: <Icon className={styles.error} type="close-circle" />,
-    success: <Icon className={styles.success} type="check-circle" />,
+    error: <Icon className={styles.error} type="close-circle" theme="filled" />,
+    success: <Icon className={styles.success} type="check-circle" theme="filled" />,
   };
   const clsString = classNames(styles.result, className);
   return (
