@@ -1,5 +1,5 @@
-import request from '../utils/request';
+import request from '@/utils/request';
 
-export async function query(code) {
+export default async function queryError(code) {
   return request(`/api/${code}`);
 }
