@@ -1,4 +1,4 @@
-import RenderAuthorized from '../components/Authorized';
+import RenderAuthorized from '@/components/Authorized';
 import { getAuthority } from './authority';
 
 let Authorized = RenderAuthorized(getAuthority()); // eslint-disable-line
