@@ -1,9 +1,19 @@
 // use localStorage to store the authority info, which might be sent from server in actual project.
 export function getAuthority() {
   // return localStorage.getItem('antd-pro-authority') || ['admin', 'user'];
-  return localStorage.getItem('antd-pro-authority') || 'admin';
+  let authority = localStorage.getItem('antd-pro-authority');
+  if (authority) {
+    if (authority.includes('[')) {
+      authority = JSON.parse(authority);
+    } else {
+      authority = [JSON.parse(authority)];
+    }
+  } else {
+    authority = ['admin'];
+  }
+  return authority;
 }
 
 export function setAuthority(authority) {
-  return localStorage.setItem('antd-pro-authority', authority);
+  return localStorage.setItem('antd-pro-authority', JSON.stringify(authority));
 }
