@@ -1,7 +1,8 @@
 import fetch from 'dva/fetch';
 import { notification } from 'antd';
-import { routerRedux } from 'dva/router';
-import store from '../index';
+import router from 'umi/router';
+import hash from 'hash.js';
+import { isAntdPro } from './utils';
 
 const codeMessage = {
   200: '服务器成功返回请求的数据。',
@@ -20,7 +21,8 @@ const codeMessage = {
   503: '服务不可用，服务器暂时过载或维护。',
   504: '网关超时。',
 };
-function checkStatus(response) {
+
+const checkStatus = response => {
   if (response.status >= 200 && response.status < 300) {
     return response;
   }
@@ -33,7 +35,26 @@ function checkStatus(response) {
   error.name = response.status;
   error.response = response;
   throw error;
-}
+};
+
+const cachedSave = (response, hashcode) => {
+  /**
+   * Clone a response data and store it in sessionStorage
+   * Does not support data other than json, Cache only json
+   */
+  const contentType = response.headers.get('Content-Type');
+  if (contentType && contentType.match(/application\/json/i)) {
+    // All data is saved as text
+    response
+      .clone()
+      .text()
+      .then(content => {
+        sessionStorage.setItem(hashcode, content);
+        sessionStorage.setItem(`${hashcode}:timestamp`, Date.now());
+      });
+  }
+  return response;
+};
 
 /**
  * Requests a URL, returning a promise.
@@ -42,7 +63,22 @@ function checkStatus(response) {
  * @param  {object} [options] The options we want to pass to "fetch"
  * @return {object}           An object containing either "data" or "err"
  */
-export default function request(url, options) {
+export default function request(
+  url,
+  options = {
+    expirys: isAntdPro(),
+  }
+) {
+  /**
+   * Produce fingerprints based on url and parameters
+   * Maybe url has the same parameters
+   */
+  const fingerprint = url + (options.body ? JSON.stringify(options.body) : '');
+  const hashcode = hash
+    .sha256()
+    .update(fingerprint)
+    .digest('hex');
+
   const defaultOptions = {
     credentials: 'include',
   };
@@ -68,33 +104,53 @@ export default function request(url, options) {
     }
   }
 
+  const expirys = options.expirys || 60;
+  // options.expirys !== false, return the cache,
+  if (options.expirys !== false) {
+    const cached = sessionStorage.getItem(hashcode);
+    const whenCached = sessionStorage.getItem(`${hashcode}:timestamp`);
+    if (cached !== null && whenCached !== null) {
+      const age = (Date.now() - whenCached) / 1000;
+      if (age < expirys) {
+        const response = new Response(new Blob([cached]));
+        return response.json();
+      }
+      sessionStorage.removeItem(hashcode);
+      sessionStorage.removeItem(`${hashcode}:timestamp`);
+    }
+  }
   return fetch(url, newOptions)
     .then(checkStatus)
+    .then(response => cachedSave(response, hashcode))
     .then(response => {
+      // DELETE and 204 do not return data by default
+      // using .json will report an error.
       if (newOptions.method === 'DELETE' || response.status === 204) {
         return response.text();
       }
       return response.json();
     })
     .catch(e => {
-      const { dispatch } = store;
       const status = e.name;
       if (status === 401) {
-        dispatch({
+        // @HACK
+        /* eslint-disable no-underscore-dangle */
+        window.g_app._store.dispatch({
           type: 'login/logout',
         });
         return;
       }
+      // environment should not be used
       if (status === 403) {
-        dispatch(routerRedux.push('/exception/403'));
+        router.push('/exception/403');
         return;
       }
       if (status <= 504 && status >= 500) {
-        dispatch(routerRedux.push('/exception/500'));
+        router.push('/exception/500');
         return;
       }
       if (status >= 404 && status < 422) {
-        dispatch(routerRedux.push('/exception/404'));
+        router.push('/exception/404');
       }
     });
 }
