@@ -3,13 +3,14 @@ const { kill } = require('cross-port-killer');
 
 const env = Object.create(process.env);
 env.BROWSER = 'none';
+env.TEST = true;
 const startServer = spawn(/^win/.test(process.platform) ? 'npm.cmd' : 'npm', ['start'], {
   env,
 });
 
 startServer.stderr.on('data', data => {
   // eslint-disable-next-line
-  console.log(data);
+  console.log(data.toString());
 });
 
 startServer.on('exit', () => {
@@ -21,10 +22,7 @@ console.log('Starting development server for e2e tests...');
 startServer.stdout.on('data', data => {
   // eslint-disable-next-line
   console.log(data.toString());
-  if (
-    data.toString().indexOf('Compiled successfully') >= 0 ||
-    data.toString().indexOf('Compiled with warnings') >= 0
-  ) {
+  if (data.toString().indexOf('App running at') >= 0) {
     // eslint-disable-next-line
     console.log('Development server is started, ready to run tests.');
     const testCmd = spawn(/^win/.test(process.platform) ? 'npm.cmd' : 'npm', ['test'], {
